+++ app.vue
@@ -31,6 +31,7 @@ useHead({
 *::before,
 *::after {
   box-sizing: border-box;
+  margin: 0;
 }
 
 html,
