--- components/icons/AddCourse.vue
@@ -1,33 +1,0 @@
-<template>
-  <svg
-    width="32"
-    height="32"
-    viewBox="0 0 32 32"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
-      d="M19.6483 3.68262H10.779C8.03366 3.68262 5.66699 5.90795 5.66699 8.65462V23.1199C5.66699 26.0213 7.87766 28.1533 10.779 28.1533H21.4297C24.1763 28.1533 26.403 25.8666 26.403 23.1199V10.7173L19.6483 3.68262Z"
-      stroke="black"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      d="M19.2981 3.66699V7.54566C19.2981 9.43899 20.8301 10.975 22.7234 10.979C24.4781 10.983 26.2741 10.9843 26.3954 10.9763"
-      stroke="black"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      opacity="0.4"
-      fill-rule="evenodd"
-      clip-rule="evenodd"
-      d="M16.5414 13.9526C16.5414 13.5384 16.2056 13.2026 15.7914 13.2026C15.3772 13.2026 15.0414 13.5384 15.0414 13.9526V16.4689H12.5244C12.1102 16.4689 11.7744 16.8047 11.7744 17.2189C11.7744 17.6331 12.1102 17.9689 12.5244 17.9689H15.0414V20.4873C15.0414 20.9015 15.3772 21.2373 15.7914 21.2373C16.2056 21.2373 16.5414 20.9015 16.5414 20.4873V17.9689H19.0591C19.4733 17.9689 19.8091 17.6331 19.8091 17.2189C19.8091 16.8047 19.4733 16.4689 19.0591 16.4689H16.5414V13.9526Z"
-      fill="black"
-    />
-  </svg>
-</template>
