--- components/icons/Course.vue
@@ -1,41 +1,0 @@
-<template>
-  <svg
-    width="32"
-    height="32"
-    viewBox="0 0 32 32"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      opacity="0.4"
-      d="M3.73877 20.636C3.73877 20.636 3.9281 22.9533 3.9721 23.684C4.03077 24.664 4.40944 25.7587 5.04144 26.5187C5.93344 27.596 6.9841 27.976 8.38677 27.9787C10.0361 27.9813 22.0294 27.9813 23.6788 27.9787C25.0814 27.976 26.1321 27.596 27.0241 26.5187C27.6561 25.7587 28.0348 24.664 28.0948 23.684C28.1374 22.9533 28.3268 20.636 28.3268 20.636"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      d="M11.3281 7.10602V6.61136C11.3281 4.98469 12.6455 3.66736 14.2721 3.66736H17.7148C19.3401 3.66736 20.6588 4.98469 20.6588 6.61136L20.6601 7.10602"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      d="M15.9933 22.2377V20.5123"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
-      d="M3.6665 11.1855V15.8082C6.22384 17.4948 9.28784 18.6762 12.6505 19.1442C13.0532 17.6762 14.3772 16.6002 15.9865 16.6002C17.5705 16.6002 18.9212 17.6762 19.2972 19.1575C22.6732 18.6895 25.7492 17.5082 28.3198 15.8082V11.1855C28.3198 8.92682 26.5025 7.10815 24.2438 7.10815H7.75584C5.49717 7.10815 3.6665 8.92682 3.6665 11.1855Z"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-  </svg>
-</template>
