--- components/icons/Criteria.vue
@@ -1,43 +1,0 @@
-<template>
-  <svg
-    width="32"
-    height="32"
-    viewBox="0 0 32 32"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      opacity="0.4"
-      d="M20.9553 21.6311H11.3286"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      opacity="0.4"
-      d="M20.9553 16.0492H11.3286"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      opacity="0.4"
-      d="M15.0019 10.4801H11.3286"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
-      d="M21.2116 3.66638C21.2116 3.66638 10.9756 3.67172 10.9596 3.67172C7.27964 3.69438 5.00098 6.11571 5.00098 9.80905V22.0704C5.00098 25.7824 7.29698 28.213 11.009 28.213C11.009 28.213 21.2436 28.209 21.261 28.209C24.941 28.1864 27.221 25.7637 27.221 22.0704V9.80905C27.221 6.09705 24.9236 3.66638 21.2116 3.66638Z"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-  </svg>
-</template>
