--- components/icons/Graduation.vue
@@ -1,43 +1,0 @@
-<template>
-  <svg
-    width="32"
-    height="32"
-    viewBox="0 0 32 32"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      opacity="0.4"
-      fill-rule="evenodd"
-      clip-rule="evenodd"
-      d="M21.779 3.66699H10.2203C6.19233 3.66699 3.66699 6.51899 3.66699 10.555V21.4457C3.66699 25.4817 6.18033 28.3337 10.2203 28.3337H21.7777C25.819 28.3337 28.3337 25.4817 28.3337 21.4457V10.555C28.3337 6.51899 25.819 3.66699 21.779 3.66699Z"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
-      d="M14.2562 16.0003C14.2562 17.363 13.1508 18.4697 11.7868 18.4697C10.4242 18.4697 9.31885 17.363 9.31885 16.0003C9.31885 14.6377 10.4242 13.531 11.7868 13.531C13.1508 13.531 14.2562 14.6377 14.2562 16.0003Z"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      d="M14.2563 16.0004H22.6803V18.4697"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      d="M18.9084 18.469V15.9996"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-  </svg>
-</template>
