--- components/icons/Lecturer.vue
@@ -1,29 +1,0 @@
-<template>
-  <svg
-    width="32"
-    height="32"
-    viewBox="0 0 32 32"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
-      d="M15.9796 20.4617C10.8228 20.4617 6.41895 21.2414 6.41895 24.3639C6.41895 27.4864 10.7948 28.2941 15.9796 28.2941C21.1364 28.2941 25.5389 27.5131 25.5389 24.3918C25.5389 21.2706 21.1643 20.4617 15.9796 20.4617Z"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      opacity="0.4"
-      fill-rule="evenodd"
-      clip-rule="evenodd"
-      d="M15.9794 16.0079C19.3635 16.0079 22.1063 13.2638 22.1063 9.87967C22.1063 6.49554 19.3635 3.75269 15.9794 3.75269C12.5952 3.75269 9.85111 6.49554 9.85111 9.87967C9.83968 13.2524 12.5648 15.9965 15.9362 16.0079H15.9794Z"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-  </svg>
-</template>
