--- components/icons/Logout.vue
@@ -1,32 +1,0 @@
-<template>
-  <svg
-    width="32"
-    height="32"
-    viewBox="0 0 32 32"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      opacity="0.4"
-      d="M20.0211 9.85265V8.60865C20.0211 5.89531 17.8211 3.69531 15.1078 3.69531H8.6078C5.8958 3.69531 3.6958 5.89531 3.6958 8.60865V23.4486C3.6958 26.162 5.8958 28.362 8.6078 28.362H15.1211C17.8265 28.362 20.0211 26.1686 20.0211 23.4633V22.206"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      d="M29.0791 16.0285H13.0244"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      d="M25.1748 12.1417L29.0788 16.0284L25.1748 19.9164"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-  </svg>
-</template>
