--- components/icons/Profile.vue
@@ -1,34 +1,0 @@
-<template>
-  <svg
-    width="32"
-    height="32"
-    viewBox="0 0 32 32"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      opacity="0.4"
-      d="M9.65967 19.7085L13.6505 14.5217L18.2028 18.0976L22.1082 13.0571"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <circle
-      cx="26.6606"
-      cy="5.60028"
-      r="2.56293"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      d="M19.8994 4.16016H10.2091C6.1938 4.16016 3.7041 7.00379 3.7041 11.019V21.7955C3.7041 25.8108 6.14499 28.6422 10.2091 28.6422H21.6812C25.6965 28.6422 28.1862 25.8108 28.1862 21.7955V12.4103"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-  </svg>
-</template>
