--- components/icons/Student.vue
@@ -1,60 +1,0 @@
-<template>
-  <svg
-    width="32"
-    height="32"
-    viewBox="0 0 32 32"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      opacity="0.4"
-      d="M23.8501 14.5289C25.7101 14.2676 27.1421 12.6729 27.1461 10.7409C27.1461 8.83693 25.7581 7.25826 23.9381 6.95959"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      opacity="0.4"
-      d="M26.3047 19.0004C28.106 19.2697 29.3634 19.9004 29.3634 21.2004C29.3634 22.095 28.7714 22.6764 27.814 23.0417"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
-      d="M15.8487 19.5518C11.5633 19.5518 7.90332 20.2011 7.90332 22.7944C7.90332 25.3864 11.5407 26.0544 15.8487 26.0544C20.134 26.0544 23.7927 25.4118 23.7927 22.8171C23.7927 20.2224 20.1567 19.5518 15.8487 19.5518Z"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
-      d="M15.8489 15.8505C18.6609 15.8505 20.9409 13.5718 20.9409 10.7585C20.9409 7.9465 18.6609 5.6665 15.8489 5.6665C13.0369 5.6665 10.7569 7.9465 10.7569 10.7585C10.7462 13.5612 13.0089 15.8412 15.8115 15.8505H15.8489Z"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      opacity="0.4"
-      d="M7.84629 14.5289C5.98496 14.2676 4.55429 12.6729 4.55029 10.7409C4.55029 8.83693 5.93829 7.25826 7.75829 6.95959"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-    <path
-      opacity="0.4"
-      d="M5.39167 19.0004C3.59034 19.2697 2.33301 19.9004 2.33301 21.2004C2.33301 22.095 2.92501 22.6764 3.88234 23.0417"
-      stroke="currentColor"
-      stroke-width="1.5"
-      stroke-linecap="round"
-      stroke-linejoin="round"
-    />
-  </svg>
-</template>
