+++ languages/en/index.js
@@ -1,7 +1,5 @@
 import seo from "./seo.json";
-import layout from "./layout.json";
 
 export const languageContents = {
-  seo: seo,
-  layout: layout,
+    seo : seo
 };
