--- languages/en/layout.json
@@ -1,10 +1,0 @@
-{
-    "courses": "Courses",
-    "students": "Students",
-    "lecturers": "Lecturers",
-    "graduation":"Graduation",
-    "prediction":"Prediction",
-    "criteria":"Criteria",
-    "profile":"Profile",
-    "logout":"Logout"
-}
