--- languages/th/layout.json
@@ -1,10 +1,0 @@
-{
-    "courses": "รายวิชา",
-    "students": "นักศึกษา",
-    "lecturers": "อาจารย์",
-    "graduation":"ประสบการณ์การศึกษา",
-    "prediction":"ตัดสินใจ",
-    "criteria":"เกณฑ์",
-    "profile":"โปรไฟล์",
-    "logout":"ออกจากระบบ"
-}
