+++ languages/th/seo.json
@@ -1,4 +1,4 @@
 {
-  "title": "Inu",
-  "desc":"Inu want to doggy style your mom"
-}
+    "welcome": "ยินดีต้อนรับ",
+    "desc":"Stating..."
+  }
