+++ nuxt.config.ts
@@ -6,9 +6,7 @@ export default defineNuxtConfig({
     "@nuxtjs/sitemap",
     "@nuxtjs/seo",
   ],
-
   tailwindcss: { exposeConfig: true },
-
   i18n: {
     defaultLocale: "en",
     strategy: "prefix_and_default",
@@ -18,21 +16,16 @@ export default defineNuxtConfig({
       { code: "th", language: "th", name: "ไทย" },
     ],
   },
-
   fonts: {
     families: [{ name: "Noto Sans Thai", fallbacks: ["Arial"], global: true }],
   },
-
   sitemap: {
     autoI18n: true,
     debug: true,
     sitemaps: false,
   },
-  devtools: false,
   seo: {
     fallbackTitle: false,
     automaticDefaults: false,
   },
-
-  compatibilityDate: "2024-10-29",
 });
