+++ package.json
@@ -25,11 +25,6 @@
     "postcss": "^8.4.47",
     "prettier": "^3.3.3",
     "sass": "^1.79.4",
-    "tailwindcss": "^3.4.13",
-    "@types/bun": "latest"
-  },
-  "module": "index.ts",
-  "peerDependencies": {
-    "typescript": "^5.0.0"
+    "tailwindcss": "^3.4.13"
   }
 }
