+++ pages/index.vue
@@ -1,14 +1,18 @@
 <template>
-  <Hero />
+  <main class="overflow-hidden text-black px-32">
+    {{ t("seo.welcome") }}
+  </main>
 </template>
 
 <script setup>
-import Hero from '@/components/course/Hero.vue';
+import { useI18n, useLocalePath, useSwitchLocalePath } from "#imports";
+
 const { t } = useI18n();
+const switchLocalePath = useSwitchLocalePath();
 
 useHead({
-  title: t("seo.title"),
-  description: t("seo.desc"),
+  title: t("seo.welcome"),
+  description: t("seo.welcome"),
 });
 
 definePageMeta({
